tests/integration/requirements.txt
# Our code is based on the AWS SDKs
boto3
botocore

# Required for gcs_backend tests
google-auth
google-cloud-storage